+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- Move from Stack/Positioned implementation to Transform.
+- BREAKING CHANGE: height and width parameters removed.  No longer necessary
+
 ## 1.0.3
 
 - Prep for publish